+++ src/components/FormValidator.js
@@ -1,12 +1,3 @@
-export const config = ({
-    formSelector: '.popup__form',
-    inputSelector: '.popup__input',
-    submitButtonSelector: '.popup__save-button',
-    inactiveButtonClass: 'popup__save-button_disabled',
-    inputErrorClass: 'popup__input_type_error',
-    errorClass: 'popup__error_visible',
-}); 
-
 export class FormValidator {
     constructor(config, form) {
         this._form = form;
