+++ src/components/PopupWithForm.js
@@ -5,10 +5,11 @@ export class PopupWithForm extends Popup {
         super(popupSelector);
         this._handleFormSubmit = handleFormSubmit;
         this._handleSubmit = this._handleSubmit.bind(this);
+        this._form = this._popup.querySelector('.popup__form');
+        this._inputsList = this._popup.querySelectorAll('.popup__input');
     }
 
     _getInputValues() {
-        this._inputsList = this._popup.querySelectorAll('.popup__input');
         this._formValues = {};
         this._inputsList.forEach((input) => {
             this._formValues[input.name] = input.value;
@@ -23,12 +24,11 @@ export class PopupWithForm extends Popup {
 
     setEventListeners() {
         super.setEventListeners();
-        this._popup.addEventListener('submit', this._handleSubmit);
+        this._form.addEventListener('submit', this._handleSubmit);
     }
 
     close() {
         super.close();
         this._form.reset();
-        this._form.removeEventListener('submit', this._handleSubmit);
     }
 }
