+++ src/index.html
@@ -68,7 +68,7 @@
             <div class="popup__container">
                 <h2 class="popup__heading">Новое место</h2>
                 <form class="popup__form popup__form_type_card" name="card-info" novalidate>
-                    <input type="text" class="popup__input" placeholder="Название" id="popupCardName" name="card-name" minlength="2" maxlength="30" required>
+                    <input type="text" class="popup__input" placeholder="Название" id="popupCardName" name="cardName" minlength="2" maxlength="30" required>
                     <span class="popup__error popupCardName-error"></span>
                     <input type="url" class="popup__input" placeholder="Ссылка на картинку" id="popupImageLink" name="link" required>
                     <span class="popup__error popupImageLink-error"></span>
